MultilanguageInvoiceExtractorApp/requirements.txt
streamlit
google_generativeai
python_dotenv
langchain
pypdf2
chromadb